pyproject.toml
[project]
name = "tasks-web"
version = "0.1.0"
description = "A task manager built with Django."
authors = [{ name = "Flurin Coretti", email = "hello@flurincoretti.com" }]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "environs[django]>=14.3.0",
    "gunicorn>=23.0.0",
    "markdown>=3.8.2",
    "psycopg2-binary>=2.9.10",
    "ulid-py>=1.1.0",
]
